Final benchmark push

+++ bench_extern.h
@@ -4,6 +4,6 @@
 #ifndef BENCH_EXTERN_H
 #define BENCH_EXTERN_H
 
-extern int bytes_allocated;
+extern long long int bytes_allocated;
 extern Dynamic_Array *times_per_node;
 #endif
